Refactor `JobOutputContents` into more generic `OutputBuffer` type
