+++ crates/brioche-core/src/utils.rs
@@ -1,5 +1,7 @@
 use std::fmt::Write as _;
 
+pub mod output_buffer;
+
 pub fn is_default<T>(value: &T) -> bool
 where
     T: Default + PartialEq,
